VANSHIKA || VSWEB-13 Changes for Blog UI
